Ignore non-existing links in links hash
In certain situations, the content_ids found in the links hash will not
exist in the publishing-api. This is correct behaviour. When this
happens, we should ignore the item.

One of the valid cases I found is collections-publisher. For any parent
topic, this puts the children in the links hash (`children` key):

https://github.com/alphagov/collections-publisher/blob/7846746aa4b5e6213
9c7e199c841c938515c18e3/app/presenters/topic_presenter.rb#L25

But these may not have been sent to the publishing-api, thus causing
this issue.