Don't rescue `Elasticsearch::DocumentNotFound`
Currently this app rescues & sends an Airbrake whenever
`Elasticsearch::DocumentNotFound` is found. This was added because I
thought that this meant that the document in question wasn't in the
index.

But there is something else going on here. At this place in the code,
`Elasticsearch::DocumentNotFound` means that the document *does* exist
in elasticsearch (it's found via
`IndexDocuments#indices_with_base_path` - which returns the index that
contains this document). But because of a bug in the
`Elasticsearch::Index` class, the Amender will still fail and raise
this error.

This commit removes the rescue so that we don't `ack` messages which we
couldn't process because of a bug. It also changes the exception
message to be slightly more accurate.

A follow up PR will be raised to fix the actual bug.