Fix amendment for non-edition docs
Amendment for docs with a type anything other than `edition` (like hmrc
manuals) currently fails because `link_to_type_and_id` doesn't
calculate the correct document type for a URL. Since there is actually
no reliable way to infer the document type, we should not be using it
at all. Instead we use the `_all` document type in the `get`, which
makes sure it returns the first document of _any_ type.