Allow misspelled ID codes to return relevant results
Often people search for ID codes (eg forms or reports) which may or may
not contain spaces or punctuation. They only see the relevant results
when they use exactly same spelling as the content, this often frustrates
users for not getting the results they are looking for.

This commit aims to solve this problem without changing too much of the
existing weight system, by applying a match query on
`all_searchable_text.id_codes` fields.