Restructure directories and modules
The elasticsearch module is a holdover from the times when we supported solr as
well. Now it is a big catch all namespace for a lot of different things.
We want to separate out indexer and searcher code, and remove the
elasticsearch namespace so we can use the official elasticsearch gem.

- Move our own elasticsearch API client to `legacy_client`

- Move deprecated advanced search functionality to `legacy_search`

- Move search functionality to `search`

- Move indexer functionality to `indexer`

- Removed the Registry namespace as this was only a single class.

- Rename unified_search -> search everywhere.

- Group together presenters for search, and workers for the indexer.

- Rename remainder of Elasticsearch namespace (dealing with index/alias
  management) to `SearchIndices`