Prevent confusion over `link` vs `document_id`
An elasticsearch document in the content indexes contains the fields
`_id`, which is the internal elasticsearch ID, and `link`, which is the
base_path or URL of the document.

In *most* cases these are interchangeable. Except for `contacts`, which
sends it's documents to rummager with an `_id` without leading slash
and `base_path` *with* the slash.

This commit renames some variables to make clear which thing we are
referring to. This makes it clear that `IndexDocuments` contains a bug
because it provides a `base_path` and not an ID.