Introduce elasticsearch gem and rake tasks for snapshot restore
- Start snapshot
- Monitor snapshot
- List snapshots in bucket
- Get latest snapshot
- Restore a named snapshot
- Monitor a restore
- Switch aliases to a restored index

The ongoing snapshot process will be managed from a cron or jenkins job,
and we plan to add a jenkins job for restoring as well.
We always create repository before snapshotting (this is idempotent), so settings
get updated when keys are rotated.

In addition to this we already have a nightly job that rebuilds the
indexes, which will be extended to clean up old snapshots.

Elasticsearch's APIs don't handle multiple snapshot/restore operations
at the same time. If we are unable to take a snapshot for this reason, we
will just let the job fail.

Ultimately, we will ensure we always have a useable snapshot by using
snapshots to refresh staging/integration and monitoring this.