Notify nicely for unknown documents
When a document is put on the queue that isn't (yet) in Rummager, we
shouldn't crash. This commit makes the processor check whether a
document from elasticsearch  is returned and raises an exception if
not.

This exception is rescued by the consumer. The plan is to use
`ProcessingError` subclasses for other types of errors too.