Boost service_manual documents by 0.3
The constants in lib/search/query_components/query.rb indicate that goverment pages are 4 times more important than service manual pages. The average booster for other document types is 1.5 so, very roughly, 0.3 makes the service_manual documents 4 times less important than the rest.

When the original service manual is retired then we can get rid of the specialisation for the service-manual index in lib/search/query_components/query.rb and elsewhere along with the actual index itself. But during the migration period we need it.