Handle malformed messages better
When deploying this branch on staging, we noticed that there are a
number of messages waiting in the queue without the `document_format`
attribute in the payload.

This commit makes sure we discard any of these messages, because there
is no way to reliably process them. `ProcessingError` do get reported
to errbit so we can take appropriate actions.