Wait for snapshot completion
When running a snapshot, pass the wait for completion flag.
The timeout needs to be increased because this can take a long time
if there are no previous snapshots (e.g. around 10 minutes for all indices)

Reverse the arguments on the check task since the second one is optional.

Added some notes about possible elasticsearch failures.