Rescue all exceptions in message queue worker
During development of the message queue processor we've experienced
multiple times that some unknown exception in the code is triggered.

Currently this will crash the worker, after which it is restarted by
the upstart. This sometimes happens very quickly, causing the message
queue processor to try the same thing many times, causing other load
troubles.

This commit makes the consumer more resilient by rescuing any errors
and waiting one second before trying again. It also reports the error
to Airbrake itself so that we have access to the offending payload.