Load entire app environment into workers
Since https://github.com/alphagov/rummager/pull/606 the sidekiq workers
have been failing because it doesn't have the `Rummager` class loaded.

This commit makes the message queue workers require the exact same code
in the message queue workers as in server-mode. This makes the code a
somewhat easier to follow, it replicates the behaviour we are used to
in Rails and prevents scenarios like this where tests are passing
because different code is loaded in test environment than in production.