Fix error reporting to Errbit
Currently the exception messages for `BulkIndexFailure` and
`FailedJobException` are constructed with a very specific error message
string. This causes Airbrake to think the errors are unique, and it
will email the admins thousands of exceptions if something goes wrong.

This commit makes the errors always have the same `message` so that is
nicely grouped in Errbit. Any additional information is sent in the
`parameters`.