Allow `content_id` to be added to documents
In the future, Rummager will be looking up the tags for an item at
indexing time. This will replace the current implementation where it
looks up content in content-api.

The publishing-api currently just exposes a unexpanded links-hash like:

```json
{
  "topics": ["bf9683e6-b584-425c-923b-dd5ed925e8d8"]
}
```

At indexing time we need to translate these content_ids into a
base_path, which is how the links are currently stored in rummager. To
ease that lookup, we will be sending the content_id of organisations,
topics and browse pages to rummager. This means we can do a simple
search in elasticsearch and use the `link` of the document.