Use Sidekiq in development
Rummager currently does not run Sidekiq in development, but has special
code paths for a synchronous and a queued mode.

This commit changes that behaviour to always use Sidekiq. A
corresponding change to the `github.gds/development` will make sure
that in development we automatically start the worker. This will keep
the development experience the same as it is now.

The main upside is that we can simplify the code and have less configs.
It also makes the development environment look more like production,
which is good.