Opt in to gds-api-adapters request tracing middleware
Pass the two headers on to all downstream requests, including elasticsearch.

We currently run a local nginx on the search hosts that does some load balancing and proxies to elasticsearch.
Long term, we can remove this, since the elasticsearch gem can do its own loadbalancing.

For now though, it would be useful to be able to trace requests back to the original request. The nginx logs
currently don't contain any identifible information about the request, because the query string is part of the
GET body rather than the URL.

We will also pass through the request id to downstream requests made through gds-api-adapters,
eg content api when indexing.