Alternative way of doing scoped search
When filtering by organisation, we expect to see the organisation page itself.
Right now, we achieve this by automatically adding the organisation page to its
own organisations field - i.e. tagging the page to itself.

Publishing api doesn't know about this tagging, and we need to make the two
consistent. This is one possible way of doing this without changing the behaviour
in rummager. The alternative is to have whitehall send organisation self-tags into
publishing api.