Fix bug in indexing
This method was caching the result of the `get_document_by_link`. This
means that after the first run, it would always re-index the same
document, causing tagging not to work for most documents.

The integration test would not have caught the issue, but I've added it
anyway because there wasn't a proper end-to-end test for this procedure.