Emit a stat to track rejected index requests
We want to count messages we receive from the rabbit
queue but choose not to process. At present the only reason
to reject index requests is if the document is not in Rummager.

This might help track down a discrepancy in the counts of the
messages hitting the rabbit queue vs the messages hitting the
sidekiq queue.

@benhyland and @davidslv