Don't rely on publishing-api when deploying
On each deploy, capistrano runs the rake task:

```
rake publishing_api:publish_special_routes
```

When publishing-api is unavailable this causes the deploy to fail.
These pages are republished each deploy for convenience, so that we
don't have to remember to republish them each time we change them. In
most cases they won't send anything new to the publishing pipeline, so
we're fine with giving a warning when the request fails.