Add Airbrake warning if unexpected links are present during indexing
Rummager now fetches links data directly from the publishing API when
indexing a document. We're in the process of updating Rummager client
apps to not send mainstream browse pages, specialist sectors and
organisation links.

This commit adds an Airbrake message if links of
this type are encountered when preparing a document for indexing.