Upgrade sidekiq to 4.1.2
There is a bug which can cause the polling thread to die
when redis is unavailable during sidekiq startup.
This means that, while the worker threads are still around,
they will never be given any work to do.

This occurs often in integration, but could potentially affect
other environments because it introduces an ordering dependency
on application startups.

The upgrade to 4.1.2 fixes the issue with the polling thread,
but introduces the explicit ordering dependency we wanted to avoid!
At least it will fail early and in a more obvious way now.

We will add a further commit to handle redis unavailability during startup.