Don't send DocumentNotFound errors to errbit
With the current design which parallelises deletes and amends, they are not avoidable.

We have better ways of seeing whether there is an unwanted difference
in content in rummager and publishing-api.

If we are worried about bugs which induce large numbers of these, we
should count them with a stat and monitor with icinga, rather than clutter errbit.