Remove notification to Airbrake in doc prepration
This was added in https://github.com/alphagov/rummager/pull/659:

> We're in the process of updating Rummager client apps to not send
mainstream browse pages, specialist sectors and organisation links.
This commit adds an Airbrake message if links of this type are
encountered when preparing a document for indexing.

We've now realised that when we do the nightly re-index, this would
send 250.000 messages to Airbrake since all documents have
`organisations`, `mainstream_browse_pages` and `specialist_sectors`.

What we _actually_ wanted to test is that clients are no longer sending
this information to the Rummager API. A replacement for that feature
would probably need to live closer to the sinatra app.