Enable scripts to skip links indexing
This commits enables us to set a flag that will prevent the document
preparer from doing links lookup in the publishing-api. It is a stopgap
while we are waiting for the publishing-api to get faster, and it
should be removed soon.

It should/must/can only be used when doing an index migration, like the
fetch-search-analytics job does[1]. In that case, it will not touch the
existing links (because `migrate_index` uses existing data).

[1]
https://github.com/alphagov/govuk-puppet/blob/master/modules/govuk_jenki
ns/templates/jobs/search_fetch_analytics_data.yaml.erb#L33-L37