Add option to log instead of raise timeouts
We're currently skipping the links-indexing during nightly reindex
(performed by the search-analytics[1] scripts).

This means that rummager links are not being synced up with the
publishing platform unless a page is republished. There are a lot
of cases where these links are known to be wrong; for example,
the old content api integration used to only append to tags, and
never remove old ones. This means there is a lot of links data
in rummager that is wrong.

Publishing API is supposed to be authoritive so we should pull as much data
as we can from it. The reason we disabled the nightly sync in the first
place is because of performance issues with the get-expanded-links call,
which affects a subset of content, depending on the amount of links
pointing to/from it. We have made some improvements but this is not
entirely fixed.

This change ensures that links are overriden from publishing api if
possible, but we fall back to the existing links otherwise.

[1]
https://github.com/alphagov/search-analytics/blob/master/nightly-run.sh