Remove snapshotting code
Pull request 578 (amended by 584, 590, 591 and 592) added the ability to create and restore Elasticsearch snapshots from Rummager.

Since this, more generic functionality has been added to govuk-puppet (pull requests 4402, 4500 and 4517) which also allows snapshotting of Elasticsearch indices.

Due to this change, this commit removes the snapshotting code added to Rummager, but keeps some refactoring that was done at the same time.