Swap out RestClient for Elasticsearch gem
https://www.elastic.co/guide/en/elasticsearch/client/ruby-api/current/_the_ruby_client.html

This gem provides a new client for communicating with the elasticsearch
cluster, Elastisearch::Transport::Client. This wraps Faraday by default which
in turn wraps Net::HTTP.

It also provides methods for each of the elasticsearch API endpoints, so we no
longer need to serialize/deserialize JSON ourselves.

Note: The client will only parse JSON responses if the response contains a
Content-Type header. We have to make sure all our stubs pass this so that our
unit tests reflect reality.

The client doesn't seem to let you specify the request Content-Type very easily
(?) so we're sending application/json for everything regardless of whether the
request has a body or not. The GdsApi::GovukHeaders headers for request tracing
have been dropped from elasticsearch requests, but we weren't using them.

Note that we no longer use Plek for finding the elasticsearch hostname - we now
need an environment variable to be passed in. This allows us to stop using an
intermediate load balancer and instead connect directly to the cluster.

Future work:

- Connect to multiple elasticsearch nodes in production
- Retry failed requests on other nodes
- Perhaps use elasticsearch-dsl to simplify query building