Swap out RestClient for Elasticsearch::Transport
This is a proof of concept to demonstrate that ripping out RestClient is
feasible.

Elasticsearch::Transport wraps Faraday by default which in turn wraps
Net::HTTP in our case.

Transport::Client is a wrapper for that with a roughly equivalent API to
LegacyClient::Client, except:

- Currently it returns the response json as a Hash, rather than an object
  with a string body.

- The exception hierarchy has been replaced by the Elasticsearch one:
https://github.com/elastic/elasticsearch-ruby/blob/master/elasticsearch-transport/lib/elasticsearch/transport/transport/errors.rb

Future work could look at look at a bigger refactoring to use the
elasticsearch-api helpers and/or more high level libraries such as
elasticsearch-dsl, instead of reimplmenting these things ourselves.

A fairly decent chunk of the code is functionality that these libraries
already provide.

Note: The client has a stupid way of handling content type headers.

The type of the response body is different depending on whether
the content type header for the response is application/json,
which means our unit tests were doing something different to
what the real code did.

Also the client doesn't seem to let you specify content type
for the request very easily (?) so we're still sending application/json
for everything regardless of whether the request has a body or not.
This means all unit tests need to include Content-Type in stubs