Alias /search.json to /unified_search.json
The public URL of the search api is /api/search.json
Currently, the internal URL is /unified_search.json

"Unified" doesn't actually mean anything anymore, it's terminology
left over from when we used to have separate searches for different
kinds of content.

gds-api-adapters only works with the internal URL because it hard codes
the "unified_search.json" part of the endpoint. Using the same name
for both internal/public URLs would avoid confusion and allow
gds-api-adapters to work with the public endpoints too.