Experiment: add hosted documentation
GitHub can host documentation under /docs. These three files are the
bare minimum to make GitHub generate a static site with Jekyll.

Not 100% sure it works, but I expect a site to be generated under:

http://alphagov.github.io/rummager/

It should contain a table of contents and link to all the docs
currently in /docs.