Fix links indexing for courts & tribunals
Content can be tagged to organisations. In most cases the organisation
has a URL in the form of

https://www.gov.uk/government/organisations/cabinet-office

Rummager internally uses a concept of `slug` to refer to other pages in
the index (in the future it should use `content_ids`, but soit).

The slug for above URL is `cabinet-office`, so the code that copies the
links from the publishing-api just strips `/government/organisations`
off the base_path. However, we didn't realise that there are
organisations with a slightly different slug, namely the courts &
tribunals, that live at:

https://www.gov.uk/courts-tribunals/administrative-court

This commit makes sure we persist the proper slug in elasticsearch for
the courts tribunals.