Add `*_content_ids` to the list of fields by default
We are changing the way we are storing and expanding information on links.
Before, there would be one fields per link type (e.g. `organisations`), which
would hold a list of `slugs` in ElasticSearch. When that document was being
served, we would expand those `slugs` into a Hash of attributes in the same
`organisations` field.

We are changing to use 2 fields per link type instead. One containing the
`content_ids` and another containing the expanded version of the links.

So, for the example above, we would have both `organisation_content_ids` and
`expanded_organisations`.

This works fine, except that if one specifically requests
`expanded_organisations` in the list of fields, we do need to include
`organisation_content_ids` as well in order to have all the necessary
information to expand those links.

This commit includes those fields by default so we are always in a position to
expand them.

If the user didn't request those `*_content_ids` fields, they won't be included
in the response.