Add support for “more like this” queries
This commit adds support in rummager for “more like this” queries to be sent to Elasticsearch. “More like this” queries allow searching using an existing indexed document as a starting point, and returning a set of documents that are “similar” to it content-wise.

A “more like this” search can be carried out using the existing rummager search endpoint with the new `similar_to` option, which takes a base path as its value (more correctly, the ID of the document in the Elasticsearch index). These searches cannot be combined with a text query (using the `q` option) and they also cannot be custom sorted (they are already sorted by similarity).