Add start and end dates to edition schema
These are used for topical events to capture their topicalness.

Currently we store this in the placeholder content schema, but not rummager
https://github.com/alphagov/govuk-content-schemas/blob/0bda3ae3e0be97cba8849eb700dfd34d9c4c3d05/dist/formats/placeholder/frontend/schema.json

This will allow us to replace /government/topical-events with a finder
of topical events, and /government/topics with a finder of policy areas,
so we have one less thing being rendered from whitehall data using a custom
template.