Rename document_type to elasticsearch_type
`document_type` is used by the content store to describe the type of a
content item. See RFC 41[1] for the background.

"document type" is also used by this application to refer to the
elasticsearch type of the document[2].

To avoid confusion and allow us to use `document_type` consistently
across GOV.UK this commit renames `document_type` to
`elasticsearch_type` where applicable.

This should not change anything for clients of this API. There are 2
todo's after this PR:

1. We currently allow clients to filter on `document_type`. This needs
to be removed or renamed to `elasticsearch_type`.
2. We currently return `document_type` in the payload by default. This
needs to be removed or renamed too.

[1]
https://gov-uk.atlassian.net/wiki/display/GOVUK/RFC+41%3A+Separate+docum
ent+type+from+format
[2]
https://www.elastic.co/guide/en/elasticsearch/guide/current/mapping.html