Add rake task to report inconsistent data
This commit adds a rake task that looks for and reports via Airbrake any facet values that are non-expanded due to unsynced data between rummager and the publishing-api.

It also adds `document_series` to the list of allowed facet fields so that it can also be tested.

`gds-api-adapters` is upgraded so that `rummager.search` can be used.