Don't allow extremely high value for `page`
We're currently seeing requests coming in via Whitehall with an
absurdly high `page` value.

It results in queries to rummager like this:

https://rummager.publishing.service.gov.uk/government/advanced_search?pa
ge=6082121121121212.1&per_page=40

Which in turn generates an elasticsearch query with a `from` value of
`243284844844848440`. This hits some kind of java integer JSON parse
limit, causing elasticsearch errors.

We have only 300.000 documents in the index, so there's absolutely no
reason to request a quintillion. 500.000 will be a good limit for the
coming years.