Catch publishing-api exceptions
This commit adds `rescue` blocks to catch timeout and HTTP exceptions raised by gds-api-adapters when communicating with the publishing-api.

Catching these exceptions allows us to log them and send them manually to Airbrake with custom messages such that errors are grouped in a sane manner.

This will prevent Airbrake seeing each message as a separate error to be reported.