Migrate from elasticsearch `facets` to `aggregates`
Facets are deprecated and stopping us migrating to newer versions of
elasticsearch.

Here we replace the internals of search to use aggregates instead of facets.
The external API can now be queried using either the old `facet_` syntax or
a new `aggregate_` syntax, with the result being nexted under a corresponding
namespace.

This code looks to do a straight like for like migration without exploring
the more advanced aggregation features which could simplify the implementation.