WIP: Remove dependence on _type field being in _source
Elasticsearch documents are uniquely identified by the combination of their `_id`
and `_type` values.

In elasticsearch 1.7, there is a `_type` metadata field, which by default is

- indexed
- not analyzed
- not stored

https://www.elastic.co/guide/en/elasticsearch/reference/1.7/mapping-type-field.html

This means you can search on it, but not retrieve it in the body of the
document.

In our configuration, we add "_type" to our mappings, so that `_type` *is*
stored in the `_source` of the document.

In elasticsearch 2.0+ this is no longer allowed; `_type` is a reserved field
name and you can't customise it anymore.

We attempted to remove this in
https://trello.com/c/mkn7SAn0/84-fix-reserved-field-name-errors-in-elasticsearch-2-x-upgrade
thinking that it wasn't needed, because search results still correctly
displayed `_type` with it missing.

It turns out that lots of rummager relies on the field being in the `_source`.
In particularly, the nightly job which rebuilds the search index, uses it
during the export.  Additionally, if the field is missing, it silently coerces
the type to edition.

The field is also used by the legacy "advanced search" result sets, although
these are only used by whitehall, and as far as I'm aware all whitehall
documents are indexed as editions.

It's also possible that rummager filters for `document_type` rely on this
source field, because they use term filters rather than type filters:
https://www.elastic.co/guide/en/elasticsearch/reference/1.7/query-dsl-type-filter.html

This commit is an attempt to remove code that relies on `_type` being in
`_source`, so that we can again try to remove the configuration from our
mappings. It also removes the dangerous behaviour of defaulting to "edition"
wherever possible. We should always preserve the original type of the document
when doing anything with already indexed documents.